refactor(providers): split stream usage normalization
